pyproject.toml
[project]
name = "reading-mate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.8",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "av>=14.2.0",
    "chroma>=0.2.0",
    "comfy-cli>=1.5.2",
    "comfyui-embedded-docs==0.3.0",
    "comfyui-frontend-package==1.28.8",
    "comfyui-workflow-templates==0.2.4",
    "dotenv>=0.9.9",
    "einops>=0.8.1",
    "fastapi>=0.120.1",
    "jupyter>=1.1.1",
    "kornia>=0.7.1",
    "langchain==0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.12",
    "langgraph==0.6.10",
    "numpy>=1.25.0",
    "openai>=2.6.0",
    "pgvector>=0.4.1",
    "pillow>=11.3.0",
    "pip>=25.3",
    "psutil>=7.1.1",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.2.11",
    "pydantic~=2.0",
    "pydantic-settings~=2.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.2",
    "pyyaml>=6.0.3",
    "safetensors>=0.4.2",
    "scipy>=1.16.3",
    "sentencepiece>=0.2.1",
    "spandrel>=0.4.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "tokenizers>=0.13.3",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchsde>=0.2.6",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.37.2",
    "uvicorn>=0.38.0",
    "yarl>=1.18.0",
]
